requirements.txt
pyserial
pandas
scikit-learn
joblib
numpy